+++ src/backend/core/models.py
@@ -922,7 +922,11 @@ class Item(TreeModel, BaseModel):
                 "brandname": settings.EMAIL_BRAND_NAME,
                 "item": self,
                 "domain": domain,
-                "link": f"{domain}/explorer/items/{self.id}/",
+                "link": (
+                    f"{domain}/explorer/items/files/{self.id}/"
+                    if self.type == ItemTypeChoices.FILE
+                    else f"{domain}/explorer/items/{self.id}/"
+                ),
                 "logo_img": settings.EMAIL_LOGO_IMG,
             }
         )